% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpal-package.R
\docType{package}
\name{cellpal-package}
\alias{cellpal-package}
\title{cellpal: Scientific colour palettes}
\description{
Publication-ready colour palettes for single-cell,
spatial transcriptomics, bulk RNA-seq, and general
scientific visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ymbouamboua/cellpal}
  \item Report bugs at \url{https://github.com/ymbouamboua/cellpal/issues}
}

}
\author{
\strong{Maintainer}: Yvon Mbouamboua \email{your.email@example.com}

}
\keyword{internal}
