% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-adjust.R
\name{cellpal_alpha}
\alias{cellpal_alpha}
\title{Set Palette Transparency}
\usage{
cellpal_alpha(palette, alpha = 1, which = NULL)
}
\arguments{
\item{palette}{Palette name, \code{cellpal} object, or character colour vector.}

\item{alpha}{Numeric transparency between \code{0} and \code{1}.}

\item{which}{Optional positions of colours to modify.}
}
\value{
A character vector with class \code{"cellpal"}.
}
\description{
Set Palette Transparency
}
\examples{
cellpal_alpha("nature", alpha = 0.5)

}
