% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-registry.R
\name{cellpal_custom_names}
\alias{cellpal_custom_names}
\title{List Custom Palettes}
\usage{
cellpal_custom_names()
}
\value{
A character vector containing custom palette names.
}
\description{
List Custom Palettes
}
