% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-accessibility.R
\name{cellpal_cvd}
\alias{cellpal_cvd}
\title{Preview a Palette under Colour-Vision Deficiencies}
\usage{
cellpal_cvd(palette, severity = 1, labels = TRUE, label_size = 2.7)
}
\arguments{
\item{palette}{Palette name, \code{cellpal} object, or colour vector.}

\item{severity}{Simulation severity between \code{0} and \code{1}.}

\item{labels}{Logical. Display hexadecimal colour labels.}

\item{label_size}{Numeric text size.}
}
\value{
A \code{ggplot2} object.
}
\description{
Preview a Palette under Colour-Vision Deficiencies
}
\examples{
\donttest{
cellpal_cvd("nature")
}

}
