% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-accessibility.R
\name{cellpal_distances}
\alias{cellpal_distances}
\title{Calculate Pairwise Palette Distances}
\usage{
cellpal_distances(palette)
}
\arguments{
\item{palette}{Palette name, \code{cellpal} object, or colour vector.}
}
\value{
A data frame containing pairwise distances.
}
\description{
Calculates pairwise Euclidean distances in CIELAB space.
}
\details{
Larger values generally indicate more distinguishable colours.
}
\examples{
cellpal_distances("nature")

}
