% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-registry.R
\name{cellpal_exists}
\alias{cellpal_exists}
\title{Check whether a Palette Exists}
\usage{
cellpal_exists(palette)
}
\arguments{
\item{palette}{Character string containing a palette name.}
}
\value{
A single logical value.
}
\description{
Check whether a Palette Exists
}
\examples{
cellpal_exists("nature")
cellpal_exists("unknown")

}
