% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-info.R
\docType{data}
\name{cellpal_info}
\alias{cellpal_info}
\title{Metadata for Built-in cellpal Palettes}
\format{
A data frame with the following columns:
\describe{
\item{name}{Palette name.}
\item{type}{Palette type: categorical, sequential, or diverging.}
\item{recommended_max}{Recommended maximum number of discrete colours.}
\item{description}{Short description of the palette.}
}
}
\usage{
cellpal_info
}
\description{
Metadata describing the built-in palettes provided by \pkg{cellpal}.
}
\keyword{datasets}
