% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{cellpal_palettes}
\alias{cellpal_palettes}
\title{Built-in cellpal palettes}
\format{
A named list of hexadecimal colour vectors.
}
\usage{
cellpal_palettes
}
\description{
A named list containing categorical, sequential, and diverging palettes.
}
\keyword{datasets}
