% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-helpers.R
\name{cellpal_seurat}
\alias{cellpal_seurat}
\title{Generate colours from Seurat metadata}
\usage{
cellpal_seurat(object, group.by = "ident", palette = "base", sort = FALSE)
}
\arguments{
\item{object}{A Seurat object.}

\item{group.by}{Metadata column or \code{"ident"}.}

\item{palette}{Palette name.}

\item{sort}{Sort category names.}
}
\value{
A named colour vector.
}
\description{
Generate colours from Seurat metadata
}
