% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-accessibility.R
\name{cellpal_simulate_cvd}
\alias{cellpal_simulate_cvd}
\title{Simulate Colour-Vision Deficiencies}
\usage{
cellpal_simulate_cvd(palette, severity = 1)
}
\arguments{
\item{palette}{Palette name, \code{cellpal} object, or colour vector.}

\item{severity}{Numeric simulation severity between \code{0} and \code{1}.}
}
\value{
A named list of colour vectors.
}
\description{
Simulates how a palette appears under deuteranopia, protanopia, and
tritanopia.
}
\examples{
cellpal_simulate_cvd("nature")

}
