% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palette.R
\name{cellpal_view}
\alias{cellpal_view}
\title{Preview a cellpal Palette}
\usage{
cellpal_view(
  palette,
  labels = TRUE,
  label_position = c("inside", "below"),
  label_size = 3,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{palette}{Palette name, a \code{cellpal} object, or a character vector of
colours.}

\item{labels}{Logical. Display hexadecimal colour labels.}

\item{label_position}{Position of labels: \code{"inside"} or \code{"below"}.}

\item{label_size}{Numeric label size.}

\item{title}{Optional plot title.}

\item{subtitle}{Optional plot subtitle.}
}
\value{
A \code{ggplot2} object.
}
\description{
Preview a cellpal Palette
}
\examples{
\donttest{
cellpal_view("nature")
cellpal_view(cellpal("heatmap_nature", n = 20, type = "continuous"))
}

}
