% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palette.R
\name{plot.cellpal}
\alias{plot.cellpal}
\title{Plot a cellpal Palette}
\usage{
\method{plot}{cellpal}(x, ...)
}
\arguments{
\item{x}{A \code{cellpal} object.}

\item{...}{Additional arguments passed to
\link[=cellpal_view]{cellpal_view()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Displays a palette as labelled colour swatches.
}
\examples{
\donttest{
plot(cellpal("nature"))
}

}
