% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-cellpal.R
\name{print.cellpal}
\alias{print.cellpal}
\title{Print a cellpal palette}
\usage{
\method{print}{cellpal}(x, ...)
}
\arguments{
\item{x}{A \code{cellpal} object.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a cellpal palette
}
