% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-cellpal.R
\name{scale_colour_cellpal}
\alias{scale_colour_cellpal}
\alias{scale_color_cellpal}
\alias{scale_fill_cellpal}
\title{cellpal colour scale for ggplot2}
\usage{
scale_colour_cellpal(palette = "nature", reverse = FALSE, ...)

scale_color_cellpal(palette = "nature", reverse = FALSE, ...)

scale_fill_cellpal(palette = "nature", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a cellpal palette.}

\item{reverse}{Logical. Reverse the palette.}

\item{...}{Arguments passed to the ggplot2 scale.}
}
\value{
A ggplot2 scale.
}
\description{
cellpal colour scale for ggplot2
}
