% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-cellpal.R
\name{scale_colour_cellpal_d}
\alias{scale_colour_cellpal_d}
\alias{scale_color_cellpal_d}
\alias{scale_fill_cellpal_d}
\title{Discrete cellpal colour scales}
\usage{
scale_colour_cellpal_d(palette = "nature", reverse = FALSE, direction = 1, ...)

scale_color_cellpal_d(palette = "nature", reverse = FALSE, direction = 1, ...)

scale_fill_cellpal_d(palette = "nature", reverse = FALSE, direction = 1, ...)
}
\arguments{
\item{palette}{Name of a cellpal palette.}

\item{reverse}{Logical. Reverse the palette.}

\item{direction}{Either \code{1} or \code{-1}.}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete cellpal colour scales
}
