% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-core.R
\name{cellpal}
\alias{cellpal}
\title{Generate a cellpal colour palette}
\usage{
cellpal(
  palette = "base",
  n = NULL,
  type = c("discrete", "continuous"),
  reverse = FALSE,
  alpha = 1,
  direction = 1
)
}
\arguments{
\item{palette}{Character or unquoted palette name.}

\item{n}{Number of colours to return. By default, returns the original
palette length.}

\item{type}{Palette mode: \code{"discrete"} or \code{"continuous"}.}

\item{reverse}{Logical. Reverse the palette.}

\item{alpha}{Numeric transparency between 0 and 1.}

\item{direction}{Integer, either \code{1} or \code{-1}. A value of \code{-1} reverses the
palette. This is an alternative to \code{reverse}.}
}
\value{
A character vector of hexadecimal colours with class \code{"cellpal"}.
}
\description{
Generate categorical or continuous palettes for single-cell,
transcriptomic, and general ggplot2 visualizations.
}
\examples{
cellpal(nature)
cellpal("jama", n = 3)
cellpal(base, n = 20)
cellpal(heatmap_nature, n = 100, type = "continuous")

}
