% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-registry.R
\name{cellpal_named}
\alias{cellpal_named}
\title{Create a Named Palette Mapping}
\usage{
cellpal_named(
  levels,
  palette = "base",
  sort = FALSE,
  reverse = FALSE,
  direction = 1
)
}
\arguments{
\item{levels}{Character or factor vector containing category names.}

\item{palette}{Name of a registered cellpal palette.}

\item{sort}{Logical. Sort categories alphabetically.}

\item{reverse}{Logical. Reverse the palette.}

\item{direction}{Either \code{1} or \code{-1}.}
}
\value{
A named \code{"cellpal"} vector.
}
\description{
Create a Named Palette Mapping
}
