% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-info.R
\name{cellpal_palette_info}
\alias{cellpal_palette_info}
\title{Retrieve Information about a cellpal Palette}
\usage{
cellpal_palette_info(palette)
}
\arguments{
\item{palette}{Character string naming a registered palette.}
}
\value{
An object of class \code{"cellpal_palette_info"}.
}
\description{
Retrieve Information about a cellpal Palette
}
\examples{
cellpal_palette_info("nature")

}
