% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-registry.R
\name{cellpal_unregister}
\alias{cellpal_unregister}
\title{Remove a Custom Palette}
\usage{
cellpal_unregister(name)
}
\arguments{
\item{name}{Name of a custom palette.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Remove a Custom Palette
}
\examples{
cellpal_register("temporary", c("#000000", "#FFFFFF"))
cellpal_unregister("temporary")

}
